zoning-kg/requirements.txt
# Core NLP
spacy>=3.7.0

# Data validation
pydantic>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Knowledge Graph
graphiti-core>=0.5.0
neo4j>=5.0.0

# OpenAI (for Graphiti LLM extraction)
openai>=1.0.0

# Testing
pytest>=7.0.0

# Progress bars
tqdm>=4.0.0

# Zep Cloud SDK
zep-cloud>=2.0.0
